Update Project With All

+++ accounts-srv/src/main/java/acme/oddo/utils/CheckValidInput.java
@@ -1,17 +1,17 @@
 package acme.oddo.utils;
 
 public class CheckValidInput {
-    
 
     public static boolean isStringNumeric(String strNum) {
         if (strNum == null) {
             return false;
         }
         try {
-            Integer.parseInt(strNum);
+            Double.parseDouble(strNum);
         } catch (NumberFormatException nfe) {
             return false;
         }
         return true;
     }
+    
 }
